sargo: Set active slot to 'b'
Sargo exposes weird behavior when flashing onto the 'a' slot.
After installation it automatically reboots into the recovery again,
causing the boot counter to decrement with every reboot due to the
phone not rebooting into a working rootfs anymore. So, the test device
wins now, set the active slot to 'b' as that's the only slot it
successfully boots from.