Add Pixel 3a (sargo) (#38)
* Add Pixel 3a (sargo)

* sargo: Also flash downloaded dtbo & vbmeta images

* sargo: Proper URLs to firmware files and recovery explanation

* sargo: Further recovery explanation

* sargo: Use phone_power_down image on bootloader screen

* sargo: Update boot.img checksum and add fastboot:set_active step

* sargo: Set active slot to 'b'

Sargo exposes weird behavior when flashing onto the 'a' slot.
After installation it automatically reboots into the recovery again,
causing the boot counter to decrement with every reboot due to the
phone not rebooting into a working rootfs anymore. So, the test device
wins now, set the active slot to 'b' as that's the only slot it
successfully boots from.