v2: devices: axolotl: set active slot to "b" for android installation
Android OS are installed via sideload, which causes to install the
actual system into the inactive slot:
  - If "a" is active, it installs to "b"
  - If "b" is active, it installs to "a"

This causes the bootloader to get confused when checking AVB and
because vbmeta_a and friends are empty, it complains about the
device being corrupt when users decide to lock the bootloader after
installation.

Set the active slot to "b" to install the new system into the "a"
slot.

Change-Id: If109c8393d27cadf24d995ab6b0df763501c9a4c
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>