devices/FP5: automatically flash stock Android 13 as part of bootstrap
Seems not every people reads the whole unlock instructions and missed
the "Android 13" requirement. So to simplify this process, just download
FP5's Android 13 factory image and flash it on user's behalf.

The image is distributed as a zip with a direct URL that we can pull
directly. A screen asking user to acknowledge Fairphone's EULA is added.

The list of partitions to flash is taken from the flashing script (also
distributed in the image zip file). As it includes some seemingly low-
level partitions, I've added a fail-safe assert to prevent flashing them
to a wrong device.